[#2] Refactor: latest version refactoring sfn
