+++ lambda/report/__init__.py
@@ -1,0 +1,1 @@
+from report.report import generate_report
