+++ lambda/ssb/checkA/__init__.py
@@ -1,3 +1,2 @@
 from ssb import text
 from ssb.ssb import checks
-from ssb.report import generate_report
