+++ lambda/ssb/checkC/__init__.py
@@ -1,0 +1,2 @@
+from ssb import text
+from ssb.ssb import checks
